+++ app/Http/Controllers/PersonsController.php
@@ -31,95 +31,111 @@ class PersonsController extends Controller
 {
     public function persons(Request $request)
     {
-        if ($request->isMethod('get')) {
-            $query = Person::query();
 
-            if ($request->has('id') && $request->get('id') != '') {
-                $query->where('id', $request->get('id'));
+        $permissions = session('user_permissions', []);
+
+        if (in_array(strtolower('show-persons'), array_map('strtolower', $permissions))) {
+            if ($request->isMethod('get')) {
+                $query = Person::query();
+
+                if ($request->has('id') && $request->get('id') != '') {
+                    $query->where('id', $request->get('id'));
+                }
+
+                if ($request->has('name') && $request->get('name') != '') {
+                    $query->where('name', 'like', '%' . $request->get('name') . '%');
+                }
+
+                if ($request->has('email') && $request->get('email') != '') {
+                    $query->whereJsonContains('emails->value', $request->get('email'));
+                }
+
+                if ($request->has('contact_number') && $request->get('contact_number') != '') {
+                    $query->whereJsonContains('contact_numbers->value', $request->get('contact_number'));
+                }
+
+                if ($request->has('organization') && $request->get('organization') != '') {
+                    $query->where('organization', $request->get('organization'));
+                }
+                $persons = $query->get();
+                $organizations = Organization::all();
+
+                return view('contacts.persons.persons', [
+                    'persons' => $persons,
+                    'organizations' => $organizations,
+                    'filters' => $request->all()
+                ]);
             }
-
-            if ($request->has('name') && $request->get('name') != '') {
-                $query->where('name', 'like', '%' . $request->get('name') . '%');
-            }
-
-            if ($request->has('email') && $request->get('email') != '') {
-                $query->whereJsonContains('emails->value', $request->get('email'));
-            }
-
-            if ($request->has('contact_number') && $request->get('contact_number') != '') {
-                $query->whereJsonContains('contact_numbers->value', $request->get('contact_number'));
-            }
-
-            if ($request->has('organization') && $request->get('organization') != '') {
-                $query->where('organization', $request->get('organization'));
-            }
-            $persons = $query->get();
-            $organizations = Organization::all();
-
-            return view('contacts.persons.persons', [
-                'persons' => $persons,
-                'organizations' => $organizations,
-                'filters' => $request->all()
-            ]);
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
         }
     }
 
     public function create_person(Request $request)
     {
-        if ($request->isMethod('get')) {
-            $organizations = Organization::get();
-            return view('contacts.persons.create_person', ['organizations' => $organizations]);
-        }
-        if ($request->isMethod('post')) {
-            $request->validate([
-                'name' => 'required|string|max:255',
-                'emails.*' => 'required|email',
-                'contact_numbers.*' => 'required|string',
-                'email_types.*' => 'required|in:work,home',
-                'number_types.*' => 'required|in:work,home',
-            ]);
-            $file_name = null;
-            if ($request->picture) {
-                $file_name = time() . '-.' . $request->picture->extension();
-                $request->picture->move(public_path('uploads/persons/pictures'), $file_name);
+
+        $permissions = session('user_permissions', []);
+
+        if (in_array(strtolower('create-persons'), array_map('strtolower', $permissions))) {
+            if ($request->isMethod('get')) {
+                $organizations = Organization::get();
+                return view('contacts.persons.create_person', ['organizations' => $organizations]);
             }
-
-            $person = new Person();
-            $person->name = $request->name;
-            $person->dob = $request->dob;
-            $person->organization = $request->organization;
-            $person->picture = $file_name;
-            if ($request->has('emails')) {
-                $emails = [];
-                foreach ($request->input('emails') as $key => $email) {
-                    $emailType = $request->input("email_types.{$key}");
-
-                    $emails[] = [
-                        'value' => $email,
-                        'label' => $emailType
-                    ];
+            if ($request->isMethod('post')) {
+                $request->validate([
+                    'name' => 'required|string|max:255',
+                    'emails.*' => 'required|email',
+                    'contact_numbers.*' => 'required|string',
+                    'email_types.*' => 'required|in:work,home',
+                    'number_types.*' => 'required|in:work,home',
+                ]);
+                $file_name = null;
+                if ($request->picture) {
+                    $file_name = time() . '-.' . $request->picture->extension();
+                    $request->picture->move(public_path('uploads/persons/pictures'), $file_name);
                 }
 
-                $person->emails = $emails;
-            }
+                $person = new Person();
+                $person->name = $request->name;
+                $person->dob = $request->dob;
+                $person->organization = $request->organization;
+                $person->picture = $file_name;
+                if ($request->has('emails')) {
+                    $emails = [];
+                    foreach ($request->input('emails') as $key => $email) {
+                        $emailType = $request->input("email_types.{$key}");
 
-            if ($request->has('contact_numbers')) {
-                $contactNumbers = [];
-                foreach ($request->input('contact_numbers') as $key => $contactNumber) {
-                    $numberType = $request->input("number_types.{$key}");
+                        $emails[] = [
+                            'value' => $email,
+                            'label' => $emailType
+                        ];
+                    }
 
-                    $contactNumbers[] = [
-                        'value' => $contactNumber,
-                        'label' => $numberType
-                    ];
+                    $person->emails = $emails;
                 }
 
-                $person->contact_numbers = $contactNumbers;
+                if ($request->has('contact_numbers')) {
+                    $contactNumbers = [];
+                    foreach ($request->input('contact_numbers') as $key => $contactNumber) {
+                        $numberType = $request->input("number_types.{$key}");
+
+                        $contactNumbers[] = [
+                            'value' => $contactNumber,
+                            'label' => $numberType
+                        ];
+                    }
+
+                    $person->contact_numbers = $contactNumbers;
+                }
+
+                $person->save();
+
+                return redirect()->back()->with('success', 'Person created successfully!');
             }
-
-            $person->save();
-
-            return redirect()->back()->with('success', 'Person created successfully!');
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
         }
     }
     // public function delete_person($id,Request $request)
@@ -133,15 +149,23 @@ class PersonsController extends Controller
     public function deleteAndAssign($personToDeleteId, $personToAssignId)
     {
 
-        // dd('Functionality to delete person with ID ' . $personToDeleteId . ' and reassign responsibilities to person with ID ' . $personToAssignId . ' is not yet implemented.');
+        $permissions = session('user_permissions', []);
 
-        Lead::where('person', $personToDeleteId)
-            ->update(['person' => $personToAssignId]);
+        if (in_array(strtolower('delete-persons'), array_map('strtolower', $permissions))) {
 
-        // Delete the old person
-        Person::where('id', $personToDeleteId)->delete();
+            // dd('Functionality to delete person with ID ' . $personToDeleteId . ' and reassign responsibilities to person with ID ' . $personToAssignId . ' is not yet implemented.');
 
-        return redirect()->back()->with('success', 'Person deleted and reassigned successfully!');
+            Lead::where('person', $personToDeleteId)
+                ->update(['person' => $personToAssignId]);
+
+            // Delete the old person
+            Person::where('id', $personToDeleteId)->delete();
+
+            return redirect()->back()->with('success', 'Person deleted and reassigned successfully!');
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
+        }
     }
 
 
@@ -149,131 +173,153 @@ class PersonsController extends Controller
 
     public function edit_person($id, Request $request)
     {
-        $person = Person::findOrFail($id);
+        $permissions = session('user_permissions', []);
 
-        if ($request->isMethod('get')) {
-            $organizations = Organization::get();
-            return view('contacts.persons.edit_person', ['person' => $person, 'organizations' => $organizations]);
-        }
-        if ($request->isMethod('post')) {
-            $request->validate([
-                'name' => 'required|string|max:255',
-                'emails.*' => 'required|email',
-                'contact_numbers.*' => 'required|string',
-                'email_types.*' => 'required|in:work,home',
-                'number_types.*' => 'required|in:work,home',
-            ]);
-            $file_name = $person->picture;
-            if ($request->picture) {
-                $file_name = time() . '-.' . $request->picture->extension();
-                $request->picture->move(public_path('uploads/persons/pictures'), $file_name);
+        if (in_array(strtolower('edit-persons'), array_map('strtolower', $permissions))) {
+            $person = Person::findOrFail($id);
+
+            if ($request->isMethod('get')) {
+                $organizations = Organization::get();
+                return view('contacts.persons.edit_person', ['person' => $person, 'organizations' => $organizations]);
             }
-            $person->name = $request->name;
-            $person->organization = $request->organization;
-            $person->dob = $request->dob;
-            $person->picture = $file_name;
-            if ($request->has('emails')) {
-                $emails = [];
-                foreach ($request->input('emails') as $key => $email) {
-                    $emailType = $request->input("email_types.{$key}");
-                    $emails[] = [
-                        'value' => $email,
-                        'label' => $emailType
-                    ];
+            if ($request->isMethod('post')) {
+                $request->validate([
+                    'name' => 'required|string|max:255',
+                    'emails.*' => 'required|email',
+                    'contact_numbers.*' => 'required|string',
+                    'email_types.*' => 'required|in:work,home',
+                    'number_types.*' => 'required|in:work,home',
+                ]);
+                $file_name = $person->picture;
+                if ($request->picture) {
+                    $file_name = time() . '-.' . $request->picture->extension();
+                    $request->picture->move(public_path('uploads/persons/pictures'), $file_name);
                 }
-                $person->emails = $emails;
-            }
-
-            if ($request->has('contact_numbers')) {
-                $contactNumbers = [];
-                foreach ($request->input('contact_numbers') as $key => $contactNumber) {
-                    $numberType = $request->input("number_types.{$key}");
-                    $contactNumbers[] = [
-                        'value' => $contactNumber,
-                        'label' => $numberType
-                    ];
+                $person->name = $request->name;
+                $person->organization = $request->organization;
+                $person->dob = $request->dob;
+                $person->picture = $file_name;
+                if ($request->has('emails')) {
+                    $emails = [];
+                    foreach ($request->input('emails') as $key => $email) {
+                        $emailType = $request->input("email_types.{$key}");
+                        $emails[] = [
+                            'value' => $email,
+                            'label' => $emailType
+                        ];
+                    }
+                    $person->emails = $emails;
                 }
-                $person->contact_numbers = $contactNumbers;
+
+                if ($request->has('contact_numbers')) {
+                    $contactNumbers = [];
+                    foreach ($request->input('contact_numbers') as $key => $contactNumber) {
+                        $numberType = $request->input("number_types.{$key}");
+                        $contactNumbers[] = [
+                            'value' => $contactNumber,
+                            'label' => $numberType
+                        ];
+                    }
+                    $person->contact_numbers = $contactNumbers;
+                }
+
+                $person->save();
+
+                return redirect()->back()->with('success', 'Person updated successfully!');
             }
-
-            $person->save();
-
-            return redirect()->back()->with('success', 'Person updated successfully!');
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
         }
     }
     public function delete_selected_persons(Request $request)
     {
-        $personIds = $request->input('selected_persons', []);
 
-        if (!empty($personIds)) {
-            Person::whereIn('id', $personIds)->delete();
-            return back()->with('success', 'Selected persons deleted successfully.');
+        $permissions = session('delete-persons', []);
+
+        if (in_array(strtolower('view-web-forms'), array_map('strtolower', $permissions))) {
+            $personIds = $request->input('selected_persons', []);
+
+            if (!empty($personIds)) {
+                Person::whereIn('id', $personIds)->delete();
+                return back()->with('success', 'Selected persons deleted successfully.');
+            }
+
+            return back()->with('error', 'No persons selected.');
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
         }
-
-        return back()->with('error', 'No persons selected.');
     }
     public function import_persons(Request $request)
     {
-        $request->validate([
-            'persons' => 'required|mimes:xls,xlsx,csv',
-        ]);
+        $permissions = session('import-persons', []);
 
-        $data = Excel::toArray([], $request->file('persons'));
-        $rows = $data[0] ?? [];
+        if (in_array(strtolower('view-web-forms'), array_map('strtolower', $permissions))) {
+            $request->validate([
+                'persons' => 'required|mimes:xls,xlsx,csv',
+            ]);
 
-        if (empty($rows)) {
-            return back()->with('error', 'Empty file');
+            $data = Excel::toArray([], $request->file('persons'));
+            $rows = $data[0] ?? [];
+
+            if (empty($rows)) {
+                return back()->with('error', 'Empty file');
+            }
+
+            $header = array_map('trim', $rows[0]); // first row: Name, Email, Number
+            array_shift($rows); // remove header
+
+            $seen = []; // track duplicates inside file
+            $inserted = 0;
+            $skipped = 0;
+
+            foreach ($rows as $row) {
+                if (empty(array_filter($row))) continue;
+
+                $record = array_combine($header, $row);
+
+                $name = trim($record['Name'] ?? '');
+                $email = strtolower(trim($record['Email'] ?? ''));
+                $number = preg_replace('/\D+/', '', $record['Number'] ?? ''); // keep only digits
+
+                if (!$email || !$number) {
+                    $skipped++;
+                    continue;
+                }
+
+                // dedupe inside the uploaded file
+                $fileKey = $email . '|' . $number;
+                if (isset($seen[$fileKey])) {
+                    $skipped++;
+                    continue;
+                }
+                $seen[$fileKey] = true;
+
+                // check DB if already exists
+                $exists = Person::whereRaw("JSON_SEARCH(emails, 'one', ?, NULL, '$[*].value') IS NOT NULL", [$email])
+                    ->whereRaw("JSON_SEARCH(contact_numbers, 'one', ?, NULL, '$[*].value') IS NOT NULL", [$number])
+                    ->exists();
+
+                if ($exists) {
+                    $skipped++;
+                    continue;
+                }
+
+                // create person
+                $person = new Person();
+                $person->name = $name;
+                $person->emails = [['value' => $email, 'label' => 'work']];
+                $person->contact_numbers = [['value' => $number, 'label' => 'work']];
+                $person->save();
+
+                $inserted++;
+            }
+
+            return back()->with('success', "Imported: $inserted, Skipped: $skipped");
+        } else {
+            // Option A: Hard stop
+            abort(403, 'Unauthorized');
         }
-
-        $header = array_map('trim', $rows[0]); // first row: Name, Email, Number
-        array_shift($rows); // remove header
-
-        $seen = []; // track duplicates inside file
-        $inserted = 0;
-        $skipped = 0;
-
-        foreach ($rows as $row) {
-            if (empty(array_filter($row))) continue;
-
-            $record = array_combine($header, $row);
-
-            $name = trim($record['Name'] ?? '');
-            $email = strtolower(trim($record['Email'] ?? ''));
-            $number = preg_replace('/\D+/', '', $record['Number'] ?? ''); // keep only digits
-
-            if (!$email || !$number) {
-                $skipped++;
-                continue;
-            }
-
-            // dedupe inside the uploaded file
-            $fileKey = $email . '|' . $number;
-            if (isset($seen[$fileKey])) {
-                $skipped++;
-                continue;
-            }
-            $seen[$fileKey] = true;
-
-            // check DB if already exists
-            $exists = Person::whereRaw("JSON_SEARCH(emails, 'one', ?, NULL, '$[*].value') IS NOT NULL", [$email])
-                ->whereRaw("JSON_SEARCH(contact_numbers, 'one', ?, NULL, '$[*].value') IS NOT NULL", [$number])
-                ->exists();
-
-            if ($exists) {
-                $skipped++;
-                continue;
-            }
-
-            // create person
-            $person = new Person();
-            $person->name = $name;
-            $person->emails = [['value' => $email, 'label' => 'work']];
-            $person->contact_numbers = [['value' => $number, 'label' => 'work']];
-            $person->save();
-
-            $inserted++;
-        }
-
-        return back()->with('success', "Imported: $inserted, Skipped: $skipped");
     }
 }
